requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.135.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-core==1.2.17
langchain-community
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.13
openai==2.26.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pinecone
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
pydantic==2.12.5
pydantic_core==2.41.5
pydantic_settings==2.12.0
pytest
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyreadline3
PyYAML==6.0.3
redis==7.2.1
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
supabase
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
unstructured
urllib3==1.26.2
uuid_utils==0.14.1
uvicorn==0.41.0
xxhash==3.6.0
zstandard==0.25.0
